address Jiabao-Sun 's review comments
